[JiShi] modify user with team development finish.

+++ src/main/kotlin/com/microservice/peopleservice/service/TeamService.kt
@@ -2,11 +2,12 @@ package com.microservice.peopleservice.service
 
 import com.microservice.peopleservice.dto.Message
 import com.microservice.peopleservice.entity.Team
+import com.microservice.peopleservice.entity.UserTeamRelation
 import com.microservice.peopleservice.repository.TeamRepository
 import com.microservice.peopleservice.repository.UserRepository
+import com.microservice.peopleservice.repository.UserTeamRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import java.util.*
 
 @Service
 class TeamService {
@@ -14,6 +15,8 @@ class TeamService {
     private lateinit var teamRepository: TeamRepository
     @Autowired
     private lateinit var userRepository: UserRepository
+    @Autowired
+    private lateinit var userTeamRepository: UserTeamRepository
 
     fun createTeam(newTeam: Team): Message {
         userRepository.findByUsername(newTeam.creator!!) ?: return Message(false, "no this user")
@@ -23,7 +26,8 @@ class TeamService {
             return Message(false, "one creator can not create same name team")
         }
 
-        teamRepository.save(newTeam)
+        val newTeamCreateSuccess = teamRepository.save(newTeam)
+        addUserIntoTeam(UserTeamRelation(newTeamCreateSuccess.creator, newTeamCreateSuccess.id))
         return Message(true, "team create success")
     }
 
@@ -31,7 +35,7 @@ class TeamService {
         userRepository.findByUsername(updateTeam.creator!!) ?: return Message(false, "no this user")
 
         val oldTeamOptional = teamRepository.findById(updateTeam.id!!)
-        if (!oldTeamOptional.isEmpty) {
+        if (oldTeamOptional.isPresent) {
             val oldTeam = oldTeamOptional.get()
             oldTeam.description = updateTeam.description
 
@@ -54,7 +58,59 @@ class TeamService {
     }
 
     fun removeTeamById(id: Int): Message {
-        teamRepository.deleteById(id)
-        return Message(true, "team remove success")
+        val teamOption = teamRepository.findById(id)
+        if (teamOption.isPresent) {
+            teamRepository.deleteById(id)
+            return Message(true, "team remove success")
+        }
+        return Message(true, "team id do not exit")
+    }
+
+    fun modifyUserWithTeamHandler(teamId: Int, username: String, add: Boolean): Message {
+        val checkMessage = checkUsernameAndTeamIdIsValidity(username, teamId)
+        if (!checkMessage.isSusses){
+            return checkMessage
+        }
+
+        val newUserTeamRelation = UserTeamRelation(username, teamId)
+        return when {
+            add -> addUserIntoTeam(newUserTeamRelation)
+            else -> removeUserOutTeam(newUserTeamRelation)
+        }
+    }
+
+    private fun checkUsernameAndTeamIdIsValidity(username: String, teamId: Int): Message {
+        userRepository.findByUsername(username) ?: return Message(false, "no this user")
+        val modifyTeamOption = teamRepository.findById(teamId)
+        return when {
+            modifyTeamOption.isEmpty -> Message(false, "team id do not exit")
+            modifyTeamOption.get().creator.equals(username) -> Message(false, "team can not modify creator by this function")
+            else -> Message(true, "ok")
+        }
+    }
+
+    private fun removeUserOutTeam(userTeamRelation: UserTeamRelation): Message {
+        return when (val userTeamRelationInDB =
+                userTeamRepository.findByUsernameAndTeamId(userTeamRelation.username!!, userTeamRelation.teamId!!)) {
+            null -> {
+                Message(false, "user do not in this team")
+            }
+            else -> {
+                userTeamRepository.delete(userTeamRelationInDB)
+                Message(true, "remove user success")
+            }
+        }
+    }
+
+    private fun addUserIntoTeam(userTeamRelation: UserTeamRelation): Message {
+        return when {
+            userTeamRepository.findByUsernameAndTeamId(userTeamRelation.username!!, userTeamRelation.teamId!!) != null -> {
+                Message(false, "user already in this team")
+            }
+            else -> {
+                userTeamRepository.save(userTeamRelation)
+                Message(true, "add user success")
+            }
+        }
     }
 }
